Replaced usernames in the links by the new project owner
For more information on the reason for this transfer:
IdefaSoft originally coded the entire project following a specific request from me. With his agreement, I posted it on this GitHub account. However, in the end, I neither had the time nor the motivation to continue the project. Given its popularity, I found it a shame to let it go to waste. So, I decided to transfer it to him, as he was the original creator of the code, and he accepted.

Wishing you all the best, and thank you for your support.